Refactor scaffold script
