+++ packages/create-sitecore-jss/src/templates/nextjs-styleguide/src/components/styleguide/Styleguide-ComponentParams.tsx
@@ -1,5 +1,6 @@
 import StyleguideSpecimen from './Styleguide-Specimen';
-import { ComponentProps, StyleguideSpecimenFields } from 'lib/component-props';
+import { ComponentProps } from 'lib/component-props';
+import { StyleguideSpecimenFields } from 'lib/component-props/styleguide';
 
 type StyleguideComponentParamsProps = ComponentProps &
   StyleguideSpecimenFields & {
