+++ packages/create-sitecore-jss/src/templates/nextjs-styleguide/src/components/styleguide/Styleguide-Layout-Tabs.tsx
@@ -1,13 +1,14 @@
 import React, { ReactElement } from 'react';
 import { withPlaceholder, withSitecoreContext, Text } from '@sitecore-jss/sitecore-jss-nextjs';
 import StyleguideSpecimen from './Styleguide-Specimen';
-import { StyleguideComponentWithContextProps, StyleguideSpecimenFields } from 'lib/component-props';
+import { ComponentWithContextProps } from 'lib/component-props';
+import { StyleguideSpecimenFields } from 'lib/component-props/styleguide';
 
 interface StyleguideLayoutTabsState {
   activeTabIndex: number;
 }
 
-type StyleguideLayoutTabsProps = StyleguideComponentWithContextProps &
+type StyleguideLayoutTabsProps = ComponentWithContextProps &
   StyleguideSpecimenFields & {
     name: string;
     tabsPlaceholder: ReactElement[];
