+++ packages/create-sitecore-jss/src/templates/nextjs-styleguide/src/lib/component-props/styleguide.ts
@@ -1,0 +1,11 @@
+import { Field } from '@sitecore-jss/sitecore-jss-nextjs';
+
+/**
+ * Shared styleguide specimen fields
+ */
+ export type StyleguideSpecimenFields = {
+  fields: {
+    description: Field<string>;
+    heading: Field<string>;
+  };
+};
